Introduce type precedence and basic type inference
Type precedence during casting enables us to decide if parameters in expressions
can be safely casted to another parameter's type. This is possible by assigning
IDs to all the types which denote the rank of the types. Higher IDs denote
higher precedence.

There are some exceptions where we allow upcasting. When the provided expression
resolves to a Literal and we wouldn't lose entropy when applying a cast, we also
allow upcasting (e.g. int to long, float to double).

The type inference is very basic but works as follows. If we have two
parameters A and B, we first try to cast A to B. This will succeed
depending on the type precedence value of the two parameters. Afterwards we try
to cast B to A. If we applied a cast in step one, we won't cast in step two;
likewise, if we didn't cast in step one we will cast in step two.

Due to these two improvements, we can now remove implicit column casting. That
means that we try to cast Literal values for columns but require an explicit
cast whenever two columns are compared which don't have a matching
type. Column casts should generally be avoided because they can slow down Lucene
queries dramatically.