Added Cluster Check for tables that need recreation.
This check is added to "catch" tables that are created before ES 5 (CrateDB < 2.0.0)
and suggest to the user to recreate them so that they will be compatible with ES 6.
The reason is that in ES 6 the backwards compatibility layer is removed, and in Lucene 7
the bwc layer is moved to solr packages, so it quite some effort will be required to re-enable it on our own.
Moreover by removing the bwc layer we will be able to use new more performant Lucene queries in some cases (eg. geo-queries).

The TablesNeedUpgradeSysCheck will inform the user to run OPTIMIZE TABLE `table_ident` WITH (upgrade_segments=true)
for all tables that are not in the current Lucene major version. Since Currently (ES 5.x) Lucene version is 6.x the
check will "catch" tables that are created before ES 5.x (Lucene 5.x) but those tables are also "cought" by the TablesNeedRecreationSysCheck.
Therefore there is no way we can have tables to trigger the TablesNeedUpgradeSysCheck and the relevant integ tests are marked
with @Ignore and a TODO is added as a reminder to renable them after the upgrade to ES 6.